Create Python_Program_to_implement_bin()_with___index__()_for_Non_Integer_Class.py
To resolve the TypeError when using the bin() method the __index__() method can be used.

In this example, we've provided an instance of the Quantity class to the bin() method. Despite Quantity() not being an integer, there's no TypeError raised. This is due to the utilization of the index() method, which ensures the return value is an integer, in this case, the sum of the fruits.