Create Python_Program_to_Transpose_a_Matrix_Using_Nested_List_Comprehension.py
The transpose of a matrix is found by interchanging its rows into columns or columns into rows. 

List comprehension is a concise way to create lists in Python. It allows you to create a new list by applying an expression to each item in an existing iterable (such as a list, tuple, or string) and optionally filtering the items based on a condition.

Syntax :
         new_list = [expression for item in iterable if condition]
-expression: The operation or calculation to be performed on each item.
-item: Variable representing each element in the iterable.
-iterable: The original iterable (e.g., list, tuple, string) you are iterating over.
-condition (optional): A filter that determines whether the item should be included in the new list.