Create Python_Program_to_Check_Whether_a_String_is_Palindrome_or_Not.py
A string is said to be palindrome if it remains the same on reading from both ends. It means that when you reverse a given string, it should be the same as the original string.

For instance, the string 'level' is a palindrome because it remains the same when you read it from the beginning to the end and vice versa.

In the program we have used the method casefold() to make it suitable for caseless comparisons. Basically, this method returns a lowercased version of the string.
Furthermore, the string is reversed using the built-in reversed() function. Since this function returns a reversed object, we use the list() function to convert them into a list before comparing.