Create Python_Program_to_Count_the_Number_of_Each_Vowel_Using_Using_List_and_Dictionary_Comprehension.py
We have nested a list comprehension inside a dictionary comprehension to count the vowels in a single line.

The dictionary comprehension runs for all vowel characters and the list comprehension inside the dictionary comprehension checks if any characters in the string match that particular vowel.

At the end, a list with 1s is generated for the number of each vowel character. The sum() method is used to calculate the sum of the elements for each list.

However, this program is slower as we iterate over the entire input string for each vowel.