Create Python_Program_to_Merge_Two_Dictionaries_Using_the_|_Operator.py
In Python, the | operator, when used with dictionaries, is called the "merge" or "update" operator. It was introduced in Python 3.9 as a shorthand for merging dictionaries. 

The | operator combines two dictionaries, and if there are common keys, the values from the second dictionary overwrite the values from the first one.