Create Python_Program_to_Merge_Two_Dictionaries_Using_the_Dictionary_Unpacking_(**_Operator).py
Dictionary unpacking is a feature in Python that allows us to merge two dictionaries or create a new dictionary by unpacking the key-value pairs from multiple dictionaries. The ** (double asterisk) is used for dictionary unpacking.

If there are overlapping keys, the values from the second dictionary overwrites the values from the first dictionary.