Create Python_Program_to_Implement_Enumerate_Function.py
The enumerate() is a built-in function that is used to iterate over a sequence (such as a list, tuple, or string) while keeping track of the index or position of the current item. 

Syntax : 
             enumerate(iterable, start=0)
- iterable: The sequence that you want to iterate over.
- start (optional): The starting value of the index. By default, it is set to 0.

The returned object is an enumerate object. An enumerate object is an iterator that produces a sequence of tuples, each containing an index and the value from the iterable.