Create Python_Program_to_Print_Colored_Text_to_the_Terminal_Using_ANSI_Escape_Sequences.py
ANSI escape codes are special sequences of characters that are used to control formatting, color, and other text output options on terminals or console windows that support ANSI standards. These codes are a standardized way to achieve text styling and coloring in a terminal.

"\x1b[38;2;5;86;243m" :
-\x1b calls a function. You can also use \033 for the same purpose.
-38;2;r;g;b helps to set RGB color. 5;86;243 are the rgb color for blue.
-m is the function name. Here, m means SGR (Select Graphics Rendition) function.

"\x1b[0m" : resets the color to the default.

Note that the actual appearance of these codes depends on the capabilities of the terminal or console being used. Not all terminals support the full range of ANSI escape codes, and the interpretation may vary between different systems and applications.