Create Python_Program_to_Print_Colored_Text_to_the_Terminal_Using_Python_Module_termcolor.py
The termcolor library is a simple Python library that provides a convenient way to print colored text to the terminal. It is often used to enhance the visual presentation of text in command-line interfaces. 
The library is not a part of the Python Standard Library, so you need to install it separately using a package manager like pip. (pip install termcolor)

The primary function is colored(text, color), where text is the string you want to colorize, and color is the desired text color. The colored function from the termcolor library allows you to specify both the text and the color. 

Note : The appearance of colored text depends on the capabilities of your terminal or console. Not all terminals support ANSI color codes, so the results may vary.