Create Python_Program_to_Represent_enum_Using_enum_Module.py
An enumeration (enum) in programming is a symbolic name for a set of unique values. Enumerations are useful for representing a fixed set of constants or options in a more readable and self-documenting way. In Python, the enum module provides a way to create enumerations.

Enumerations are created either by using class syntax, or by using function-call syntax.