Create Python_Program_to_Find_Hash_of_File.py
Hash functions take an arbitrary amount of data and return a fixed-length bit string. The output of the function is called the digest message.

They are widely used in cryptography for authentication purposes. There are many hashing functions like MD5, SHA-1 etc. 

In this example, we will illustrate how to hash a file. We will use the SHA-1 hashing algorithm. The digest of SHA-1 is 160 bits long.

We do not feed the data from the file all at once, because some files are very large to fit in memory all at once. Breaking the file into small chunks will make the process memory efficient.